Pack the float stream in native order, as the loop it replaces does
`_object_stream`'s float branch cast to an explicit `<f8` while `_update_scalar`
packs `struct.pack("=d", value)` and its array fallback uses `np.float64`, both
native order. The two agree on every little-endian host -- which is why the
byte-parity tests pass here -- and disagree on any other, so the helper's
"byte-for-byte identical output is the contract" would not hold on a big-endian
build. The cast is now `np.float64`.

The length prefixes stay explicitly little-endian on purpose: the loop writes
them with `int.to_bytes(8, "little")`, so `_framed`'s `<u8` and the integer
branch's `<u8` lengths are the correct match. The comment now says which is
which. Byte-identity tests unchanged: 34 passed.

+++ packages/microcosm-graph/src/microcosm/graph/executor.py
@@ -489,8 +489,13 @@ def _object_stream(series: pd.Series) -> bytes | None:
     if dtype.kind == "f":
         # Every float width boxes to an exact Python float, which packs to the
         # same eight native-order IEEE-754 bytes as the float64 cast, NaN
-        # payload, signed zero and infinities included.
-        wide = np.ascontiguousarray(values, dtype="<f8")
+        # payload, signed zero and infinities included. The cast is to
+        # `np.float64`, whose byte order is the host's, because the loop this
+        # replaces packs `struct.pack("=d", value)`: an explicit `<f8` would
+        # agree with it on every little-endian host and disagree on any other.
+        # The length prefixes stay explicitly little-endian, because the loop
+        # writes them with `int.to_bytes(8, "little")`.
+        wide = np.ascontiguousarray(values, dtype=np.float64)
         payloads = np.empty((rows, 9), dtype=np.uint8)
         payloads[:, 0] = ord("f")
         payloads[:, 1:] = np.frombuffer(wide.tobytes(), dtype=np.uint8).reshape(rows, 8)
